- Extracted `#call_with` to its own module, without any changes to it.
- Extracted + Refactored `.make_validator` in its own module.
- Add ability to override validator strategies and add new ones (sort of, for user custom contracts)
- Make all calls to `Contract.make_validator` faster by memoizing

fixes #150 
fixes #156 

Example code for overriding `:class` strategy:

```ruby
require "rspec"
require "contracts"

Contract.override_validator(:class) do |contract|
  lambda do |arg|
    arg.is_a?(contract) || arg.is_a?(RSpec::Mocks::Double)
  end
end
```

@robnormal What do you think, can you use this kind of overriding mechanism?

/cc @egonSchiele Can you take a look at this one?