@egonSchiele 2015-06-09
Neat! This is much better. Did you notice any speed difference with this change?