@waterlink 2015-06-09
There is no speed difference with our current benchmarks. But given a thought to it, I think this may cause a slowdown for `Hash` contract. I think I will add this benchmark here