@waterlink 2015-06-09
I have pushed benchmark itself right into master (no-op code).

`master`:

```
                                     user     system      total        real
testing add                      1.350000   0.000000   1.350000 (  1.360404)
testing contracts add           12.650000   0.000000  12.650000 ( 12.683449)
MethodProfiler results for: Contract
+------------------------+----------+----------+--------------+-------------+-------------+
| Method                 | Min Time | Max Time | Average Time | Total Time  | Total Calls |
+------------------------+----------+----------+--------------+-------------+-------------+
| #call_with             | 0.598 ms | 7.528 ms | 0.742 ms     | 7418.086 ms | 10000       |
| .valid?                | 0.048 ms | 6.801 ms | 0.069 ms     | 1376.411 ms | 20000       |
| #maybe_append_options! | 0.047 ms | 0.685 ms | 0.064 ms     | 642.044 ms  | 10000       |
| .make_validator        | 0.004 ms | 6.179 ms | 0.007 ms     | 138.552 ms  | 20000       |
| #failure_exception     | 0.003 ms | 0.555 ms | 0.005 ms     | 54.535 ms   | 10000       |
| #method                | 0.003 ms | 0.615 ms | 0.005 ms     | 90.243 ms   | 20000       |
| #ret_contract          | 0.003 ms | 0.512 ms | 0.004 ms     | 44.013 ms   | 10000       |
| #maybe_append_block!   | 0.003 ms | 0.162 ms | 0.004 ms     | 40.909 ms   | 10000       |
| #args_contracts        | 0.003 ms | 0.182 ms | 0.004 ms     | 80.797 ms   | 20000       |
+------------------------+----------+----------+--------------+-------------+-------------+
MethodProfiler results for: Object
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
| Method                                              | Min Time | Max Time | Average Time | Total Time  | Total Calls |
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
| #contracts_add                                      | 0.710 ms | 7.652 ms | 0.872 ms     | 8717.126 ms | 10000       |
| #__contracts_ruby_original_contracts_add_iapwejs... | 0.004 ms | 0.782 ms | 0.006 ms     | 55.634 ms   | 10000       |
| .__contracts_engine                                 | 0.004 ms | 0.037 ms | 0.005 ms     | 0.484 ms    | 97          |
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
MethodProfiler results for: Contracts::MethodDecorators
+-------------------------+----------+----------+--------------+------------+-------------+
| Method                  | Min Time | Max Time | Average Time | Total Time | Total Calls |
+-------------------------+----------+----------+--------------+------------+-------------+
| #method_added           | 0.066 ms | 0.600 ms | 0.144 ms     | 6.044 ms   | 42          |
| #singleton_method_added | 0.065 ms | 0.111 ms | 0.087 ms     | 0.786 ms   | 9           |
+-------------------------+----------+----------+--------------+------------+-------------+
MethodProfiler results for: Contracts::Decorator
MethodProfiler results for: Contracts::Support
+----------------+----------+----------+--------------+------------+-------------+
| Method         | Min Time | Max Time | Average Time | Total Time | Total Calls |
+----------------+----------+----------+--------------+------------+-------------+
| .eigenclass_of | 0.004 ms | 0.034 ms | 0.006 ms     | 0.225 ms   | 39          |
+----------------+----------+----------+--------------+------------+-------------+
MethodProfiler results for: UnboundMethod
```

This branch:

```
                                     user     system      total        real
testing add                      1.360000   0.000000   1.360000 (  1.365564)
testing contracts add           14.880000   0.000000  14.880000 ( 14.929929)
MethodProfiler results for: Contract
+------------------------+----------+----------+--------------+------------+-------------+
| Method                 | Min Time | Max Time | Average Time | Total Time | Total Calls |
+------------------------+----------+----------+--------------+------------+-------------+
| #maybe_append_options! | 0.047 ms | 0.686 ms | 0.059 ms     | 590.857 ms | 10000       |
| .valid?                | 0.007 ms | 0.555 ms | 0.009 ms     | 183.756 ms | 20000       |
| #failure_exception     | 0.003 ms | 0.526 ms | 0.005 ms     | 45.985 ms  | 10000       |
| #maybe_append_block!   | 0.003 ms | 0.813 ms | 0.004 ms     | 44.937 ms  | 10000       |
| #method                | 0.003 ms | 0.583 ms | 0.004 ms     | 77.914 ms  | 20000       |
| #ret_contract          | 0.003 ms | 0.518 ms | 0.004 ms     | 38.197 ms  | 10000       |
| #args_contracts        | 0.003 ms | 0.569 ms | 0.004 ms     | 74.744 ms  | 20000       |
+------------------------+----------+----------+--------------+------------+-------------+
MethodProfiler results for: Object
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
| Method                                              | Min Time | Max Time | Average Time | Total Time  | Total Calls |
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
| #contracts_add                                      | 0.551 ms | 6.971 ms | 0.637 ms     | 6365.304 ms | 10000       |
| #__contracts_ruby_original_contracts_add_iapwhm8... | 0.004 ms | 0.526 ms | 0.005 ms     | 47.899 ms   | 10000       |
| .__contracts_engine                                 | 0.004 ms | 0.012 ms | 0.005 ms     | 0.457 ms    | 97          |
+-----------------------------------------------------+----------+----------+--------------+-------------+-------------+
MethodProfiler results for: Contracts::MethodDecorators
+-------------------------+----------+----------+--------------+------------+-------------+
| Method                  | Min Time | Max Time | Average Time | Total Time | Total Calls |
+-------------------------+----------+----------+--------------+------------+-------------+
| #method_added           | 0.062 ms | 0.694 ms | 0.142 ms     | 5.946 ms   | 42          |
| #singleton_method_added | 0.065 ms | 0.119 ms | 0.085 ms     | 0.762 ms   | 9           |
+-------------------------+----------+----------+--------------+------------+-------------+
MethodProfiler results for: Contracts::Decorator
MethodProfiler results for: Contracts::Support
+----------------+----------+----------+--------------+------------+-------------+
| Method         | Min Time | Max Time | Average Time | Total Time | Total Calls |
+----------------+----------+----------+--------------+------------+-------------+
| .eigenclass_of | 0.004 ms | 0.038 ms | 0.005 ms     | 0.194 ms   | 39          |
+----------------+----------+----------+--------------+------------+-------------+
MethodProfiler results for: UnboundMethod
```

Yes, as expected there is a bit of slowdown here. All contracts that call to `Contract.valid?` under the hood are affected by this.
I have one idea, how to optimize, I will try it out and post the results.